feat(AVO-3052): add email pref. page that logged out users can visit
https://meemoo.atlassian.net/browse/AVO-3052
