+++ src/settings/components/Email/hooks/getEmailPreferences.ts
@@ -1,0 +1,21 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { QUERY_KEYS } from '../../../../shared/constants/query-keys';
+import { CampaignMonitorService } from '../../../../shared/services/campaign-monitor-service';
+
+export const useGetEmailPreferences = (
+	preferenceCenterKey: string,
+	options: {
+		enabled: boolean;
+	} = {
+		enabled: true,
+	}
+) => {
+	return useQuery(
+		[QUERY_KEYS.GET_EMAIL_PREFERENCES, preferenceCenterKey],
+		() => {
+			return CampaignMonitorService.fetchNewsletterPreferences(preferenceCenterKey);
+		},
+		options
+	);
+};
