+++ src/settings/components/index.ts
@@ -1,4 +1,4 @@
 export { default as Account } from '../components/Account';
-export { default as Email } from '../components/Email';
+export { default as Email } from '../components/Email/Email';
 export { default as Notifications } from '../components/Notifications';
 export { default as Profile } from '../components/Profile';
