+++ src/shared/constants/query-keys.ts
@@ -7,4 +7,5 @@ export enum QUERY_KEYS {
 	GET_INTERACTIVE_TOUR_FOR_PAGE = 'GET_INTERACTIVE_TOUR_FOR_PAGE',
 	GET_WORKSPACE_COUNTS = 'GET_WORKSPACE_COUNTS',
 	GET_ITEM_USED_BY = 'GET_ITEM_USED_BY',
+	GET_EMAIL_PREFERENCES = 'GET_EMAIL_PREFERENCES',
 }
