+++ src/shared/services/campaign-monitor-service.ts
@@ -1,6 +1,6 @@
 import { fetchWithLogoutJson } from '@meemoo/admin-core-ui';
 import { type Avo } from '@viaa/avo2-types';
-import queryString from 'query-string';
+import { stringifyUrl } from 'query-string';
 
 import { CustomError, getEnv } from '../helpers';
 
@@ -8,28 +8,30 @@ export type EmailTemplateType = 'item' | 'collection' | 'bundle';
 
 export class CampaignMonitorService {
 	public static async fetchNewsletterPreferences(
-		email: string
+		preferenceCenterKey?: string
 	): Promise<Avo.Newsletter.Preferences> {
 		try {
 			return fetchWithLogoutJson<Avo.Newsletter.Preferences>(
-				`${getEnv('PROXY_URL')}/campaign-monitor/preferences?${queryString.stringify({
-					email,
-				})}`
+				stringifyUrl({
+					url: `${getEnv('PROXY_URL')}/campaign-monitor/preferences`,
+					query: {
+						preferenceCenterKey,
+					},
+				})
 			);
 		} catch (err) {
-			throw new CustomError('Failed to fetch newsletter preferences', err, {
-				email,
-			});
+			throw new CustomError('Failed to fetch newsletter preferences', err);
 		}
 	}
 
 	public static async updateNewsletterPreferences(
-		preferences: Partial<Avo.Newsletter.Preferences>
+		preferences: Partial<Avo.Newsletter.Preferences>,
+		preferenceCenterKey?: string
 	) {
 		try {
 			await fetchWithLogoutJson(`${getEnv('PROXY_URL')}/campaign-monitor/preferences`, {
 				method: 'POST',
-				body: JSON.stringify({ preferences }),
+				body: JSON.stringify({ preferences, preferenceCenterKey }),
 			});
 		} catch (err) {
 			throw new CustomError('Failed to update newsletter preferences', err, {
