+++ src/shared/translations/nl.json
@@ -1632,13 +1632,15 @@
   "settings/components/account___wachtwoord": "Wachtwoord",
   "settings/components/account___wijzig-gegevens": "Wijzig mijn gegevens",
   "settings/components/account___wijzig-wachtwoord": "Wijzig wachtwoord",
+  "settings/components/email___controleer-of-de-url-correct-is-en-contacteer-indien-nodig-de-helpdesk": "Controleer of de url correct is, en contacteer indien nodig de helpdesk",
   "settings/components/email___de-nieuwsbriefvoorkeuren-konden-niet-worden-geupdatet": "Je e-mailvoorkeuren konden niet worden bijgewerkt. Contacteer ons via de Feedback-knop mocht dit probleem zich blijven voordoen.",
-  "settings/components/email___de-nieuwsbriefvoorkeuren-konden-niet-worden-opgevraagd": "Je e-mailvoorkeuren konden niet worden opgevraagd. Contacteer ons via de Feedback-knop mocht dit probleem zich blijven voordoen.",
   "settings/components/email___e-mail-nieuwsbrief-voorkeuren": "Jouw e-mailvoorkeuren",
   "settings/components/email___e-mail-nieuwsbrief-voorkeuren-beschrijving": "Tips en inspiratie voor je lessen, workshops, vacatures, nieuws van de partners waarmee we werken aan beeld en geluid op maat van jouw klas, en dit recht in je mailbox? Schrijf je hieronder in voor onze nieuwsbrief en andere e-mailcommunicatie. Je kan je voorkeuren altijd aanpassen.",
+  "settings/components/email___het-ophalen-van-je-email-voorkeuren-is-mislukt": "Het ophalen van je email voorkeuren is mislukt",
   "settings/components/email___je-voorkeuren-zijn-opgeslagen": "Je voorkeuren zijn opgeslagen",
   "settings/components/email___nieuwsbrief-voorkeuren-pagina-beschrijving": "Duid aan welke e-mailcommunicatie je van ons wil ontvangen.",
   "settings/components/email___nieuwsbrief-voorkeuren-pagina-titel": "E-mailvoorkeuren",
+  "settings/components/email___probeer-uit-te-loggen-en-opnieuw-in-te-loggen-als-het-probleem-zich-blijft-voordoen-contacteer-dan-de-helpdesk": "Probeer uit te loggen en opnieuw in te loggen. Als het probleem zich blijft voordoen, contacteer dan de helpdesk",
   "settings/components/linked-accounts___aanmelden-met-je-leer-id": "Aanmelden met je LeerID",
   "settings/components/linked-accounts___ben-je-zeker-dat-je-de-account-koppeling-wil-verbreken": "Ben je zeker dat je de account koppeling wil verbreken?",
   "settings/components/linked-accounts___burgerprofiel": "Burgerprofiel",
